requirements.txt
-i https://pypi.org/simple
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
discord-py==1.2.4
idna==2.8
multidict==4.5.2
pbwrap==1.1.0
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.6
websockets==6.0
yarl==1.3.0
